requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow>=2.10.0
geemap
earthengine-api
rasterio
scikit-learn
folium
python-dotenv
scipy
streamlit 
streamlit_folium 
scipy.ndimage
geopy.geocoders